L1 regularization for fista
we implement the L1 regularization sovler for FISTA. so it can solve the
unconstrained minimization problem with the form of a convex smooth part
plus a L1 norm of the variable. use with_L1_reg=True to enable it.

based on above, we implement the Logistic Regression with L1
regularization model.

furthermore, a simple unittest file is added also.