bug fix and update L select strategy
objective(y) is incorrectly used as f(y) (f is the convex smooth part of
objective F), which resulted in adding an extra L1 norm of y wrongly.

And let L decrease by a factor of ETA when the condition F(pl(y)) <=
G(pl(y),y) is still satisfied.